Added `get_screenshot` method
Used file pointers for size in `get_screenshot`

Swapped back to `getsize` in `get_screenshot`

`getsize` returns different values for Windows and Linux systems because of the binary/proper form differences between bytes to kb and so on. Sticking with `getesize` for now and updating the tests to include both possible file sizes, but a better test should be created for this method.

Update test_all.py